% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-functions.R
\name{contingencyTable}
\alias{contingencyTable}
\title{Create contingency Table}
\usage{
contingencyTable(x, cols, signatures, features = "rownames")
}
\arguments{
\item{x}{A SummarizedExperiment with rowData or a data frame.}

\item{cols}{A character vector containing the names of the columns (must be logical) with the target (TRUE) and reference sets (FALSE).}

\item{signatures}{A named list with signature sets.}

\item{features}{A character vector of length 1. In case that \code{x} is a SummarizedExperiment, this argument will be
used to create the features column. In case that \code{x} is a data frame, this argument must indicate
the name of the column containing the features names (taxa, genes, etc.).}
}
\value{
A list of 2x2 contingenc tables (xtabs class).
}
\description{
\code{contingencyTable} creates a list of 2x2 contingency tables for Fisher's exact test.
}
\examples{
library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
contabs <- contingencyTable(x = se, cols = c("up", "down"), signatures = sigs)
contabs

df <- as.data.frame(SummarizedExperiment::rowData(se))
contabs2 <- contingencyTable(x = df, features = "taxa", cols = c("up", "down"), signatures = sigs)


}
\seealso{
\code{\link{contingencyTable}};
\code{\link{enrichmentTestFisher}};
\code{\link{ora}}

Other enrichment functions: 
\code{\link{enrichmentTestFisher}()},
\code{\link{ora}()}
}
\concept{enrichment functions}
