% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-functions.R
\name{enrichmentTestFisher}
\alias{enrichmentTestFisher}
\title{Enrichment with Fisher's Exact Test (Hypergeometric)}
\usage{
enrichmentTestFisher(contabs)
}
\arguments{
\item{contabs}{A iist of contingency tables (xtabs objects); output from \code{\link{contingencyTable}}.}
}
\value{
A data frame; a report containing the p-value and odds ratio for each set x signature combination
in the input list  of contingency tables
}
\description{
\code{enrichmentTestFisher} performs a one-sided overrepresentation enrichment analysis
using Fisher's exact test (hypergeometric). The function also calculates the odds.ratio
using the \code{epitools::oddsratio.wald} function.
}
\examples{
library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
contabs <- contingencyTable(x = se, cols = c("up", "down"), signatures = sigs)
contabs

enrichment_result <- enrichmentTestFisher(contabs)
enrichment_result

}
\seealso{
\code{\link{contingencyTable}};
\code{\link{enrichmentTestFisher}};
\code{\link{ora}}

Other enrichment functions: 
\code{\link{contingencyTable}()},
\code{\link{ora}()}
}
\concept{enrichment functions}
