% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleSE}
\alias{exampleSE}
\title{Example SummarizedExperment}
\usage{
exampleSE()
}
\value{
A SummarizedExperiment.
}
\description{
\code{\link{exampleSE}} creates an example of SummarizedExperiment (toy data).
}
\examples{

library(misctoolsr)
se <- exampleSE()
se

}
