% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleSignatures}
\alias{exampleSignatures}
\title{Example Signatures}
\usage{
exampleSignatures()
}
\value{
A named list of signatures.
}
\description{
\code{\link{exampleSignatures}} creates an example of signatures (toy data).
}
\examples{
library(misctoolsr)
sigs <- exampleSignatures()

}
