% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawLinksFromGithubDir.R
\name{getRawLinksFromGithubDir}
\alias{getRawLinksFromGithubDir}
\title{Get raw links from GitHub directory}
\usage{
getRawLinksFromGithubDir(repo, dir, ext = ".*")
}
\arguments{
\item{repo}{Character string; user and repo names.}

\item{dir}{Character string; directory name.}

\item{ext}{Character string; the extension of the files. Default: ".*"
(every file in the directory).}
}
\value{
A character vector with the raw links of all the files.
}
\description{
/code{getRawLinksFromGithubDir} get the list of links leading to raw files
contained in a GitHub repository (recursively).
}
\examples{
my_repo <- "bovee/fattyacids"
my_dir <- "reference_data"
my_ext <- "tsv"
raw_links <- getRawLinksFromGithubDir(
    repo = my_repo, dir = my_dir, ext = my_ext
)
}
