% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-functions.R
\name{ora}
\alias{ora}
\title{Overrepresentation Analysis}
\usage{
ora(x, cols, signatures, features = "rownames", alpha = 0.05)
}
\arguments{
\item{x}{A SummarizedExperiment with rowData or a data frame.}

\item{cols}{A character vector containing the names of the columns (must be logical) with the target (TRUE) and reference sets (FALSE).}

\item{signatures}{A named list with signature sets.}

\item{features}{A character vector of length 1. In case that \code{x} is a SummarizedExperiment, this argument will be
used to create the features column. In case that \code{x} is a data frame, this argument must indicate
the name of the column containing the features names (taxa, genes, etc.).}

\item{alpha}{Minimum p-value to filter results. Defult is 0.05.}
}
\value{
A data frame; a report containing the p-value and odds ratio for each set x signature combination
in the input list  of contingency tables
}
\description{
\code{ora} performs an overrepresentation enrichment analysis (Fisher's test; hypergeometric)
given a data set and a list of signatures. This function is a wrapper around
\code{\link{contingencyTable}} and \code{\link{enrichmentTestFisher}}.
}
\examples{
library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
ora_results <- ora(x = se, cols = c("up", "down"), signatures = sigs)
}
\seealso{
\code{\link{contingencyTable}};
\code{\link{enrichmentTestFisher}};
\code{\link{ora}}

Other enrichment functions: 
\code{\link{contingencyTable}()},
\code{\link{enrichmentTestFisher}()}
}
\concept{enrichment functions}
