% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-functions.R
\name{plotFisherRes}
\alias{plotFisherRes}
\title{Plont Enrichment Results}
\usage{
plotFisherRes(x, alpha = 0.05)
}
\arguments{
\item{x}{Output from \code{enrichmentTestFisher} or \code{ora}}

\item{alpha}{Significance threshold; default is 0.05.}
}
\value{
A plot (ggplot2) object.
}
\description{
\code{plotFisherRes} plots the output of calling either \code{\link{enrichmentTestFisher}} or
\code{\link{ora}}.
}
\examples{

library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
contabs <- contingencyTable(x = se, cols = c("up", "down"), signatures = sigs)
enrichment_result <- enrichmentTestFisher(contabs)
enrichment_plot <- plotFisherRes(enrichment_result)
enrichment_plot


}
